services/src/pyproject.toml
[tool.poetry]
name = "be_api"
version = "0.1.0"
description = "service to manage be logic"
authors = ["nthedao <nthedao2705@gmail.com>"]
license = "private"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
pep8-naming = "^0.12.1"
httpx = "~0.18.1"
python-dotenv = "^0.20.0"
"stomp.py" = "^8.0.1"
psycopg2-binary = "^2.9.3"
loguru = "^0.6.0"
SQLAlchemy = "^1.4.36"
gunicorn = "^20.1.0"
tenacity = "^8.0.1"
alembic = "^1.7.7"
autopep8 = "^1.6.0"
pymssql = "2.1.5"
pyodbc = "^4.0.34"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
bandit = "^1.7.4"
safety = "^1.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
